Support for sensor power mode and sensor delay
Add support for "SensorPowerMode" and "SensorDelay",
":CHAN%d:SENS:MOD" and ":CHAN%d:SENS:DEL".

The power mode has 3 values:
0 == disabled,
1 == enabled,
2 == power save.
When disabled, the controller will report error 260.
Since we read the sensor power mode and react on it
by setting sensorIsDisabled_ to 1 we can ignore this error.

It may happen that the device is configured by a vendor tool
outside EPICS.
To be able to follow what is inside tha controller:
Add the reading of sensor power mode and sensor delay in poll();

Improve report():
 - decode STAT in report()
 - print sensorIsDisabld

An observation:
Reading the position via ':CHAN0:POS?' runs into a timeout when
the sensor is disabled. So don't do that.