Remove 64 bit integert readback
Since the numerical range of positions is more than what a 32 bit
value can hold, the controller uses 64 bit (signed) integers.
However, we don't need the full 64 bit range.
For the real existing hardware we can use a double.
So cleanup and remove the SMARACT_ASYN_ASYNPARAMINT64 and all code
around "ireadback" and keep "freadback" as is.