smarActMCS2MotorDriver.cpp: Calculate theoretical postion in step move
When there is no enabled sensor and a move is done by
commanding n steps try to calculate the theoretical position
according to the direction and step size forward/backward.

Site note: This makes motorPosition jump at the beginning
of the movement. Should be better at the end.

Add some error handling as well.