smarActMCS2MotorDriver: Add MotorPosWhenDone
Add a helper variable which is updated AFTER the move:
When running open loop without sensor, the motor position
is calculated according to the step size.
And this value is put into motorPosition before the move.
This may confuse users/3rd party systems.
Update the new PV after the move is done.

Changes to be committed:
    modified:   smarActApp/src/smarActMCS2MotorDriver.cpp
    modified:   smarActApp/src/smarActMCS2MotorDriver.h